add documentation and structure to virtual cube tutorials Signed-off-by: dbulahov <bulahovdenis@gmail.com>
