+++ instance/emf/tutorial/virtualcube/calculatedmember/pom.xml
@@ -10,23 +10,12 @@
 * SPDX-License-Identifier: EPL-2.0
 **********************************************************************/
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.eclipse.daanse</groupId>
-    <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial</artifactId>
+    <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.virtualcube</artifactId>
     <version>${revision}</version>
   </parent>
-  <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.virtualcube</artifactId>
-  <packaging>pom</packaging>
-
-  <modules>
-    <module>min</module>
-    <module>calculatedmember</module>
-    <module>dimensions</module>
-    <module>unvisiblereferencecubes</module>
-  </modules>
-
+  <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.virtualcube.calculatedmember</artifactId>
 </project>
