+++ instance/emf/tutorial/virtualcube/calculatedmember/src/main/java/org/eclipse/daanse/rolap/mapping/instance/emf/tutorial/virtualcube/calculatedmember/CatalogSupplier.java
@@ -1,0 +1,223 @@
+/*
+ * Copyright (c) 2025 Contributors to the Eclipse Foundation.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *
+ */
+package org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.virtualcube.calculatedmember;
+
+import static org.eclipse.daanse.rolap.mapping.emf.rolapmapping.provider.util.DocumentationUtil.document;
+
+import java.util.List;
+
+import org.eclipse.daanse.rolap.mapping.api.CatalogMappingSupplier;
+import org.eclipse.daanse.rolap.mapping.api.model.CatalogMapping;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.CalculatedMember;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Catalog;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Column;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.ColumnType;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.DatabaseSchema;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.DimensionConnector;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Documentation;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.ExplicitHierarchy;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Level;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.MeasureGroup;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.PhysicalCube;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.PhysicalTable;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.RolapMappingFactory;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.StandardDimension;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.SumMeasure;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.TableQuery;
+import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.VirtualCube;
+import org.eclipse.daanse.rolap.mapping.instance.api.Kind;
+import org.eclipse.daanse.rolap.mapping.instance.api.MappingInstance;
+import org.eclipse.daanse.rolap.mapping.instance.api.Source;
+import org.osgi.service.component.annotations.Component;
+
+@Component(service = CatalogMappingSupplier.class)
+@MappingInstance(kind = Kind.TUTORIAL, number = "2.15.3", source = Source.EMF, group = "VirtualCube") // NOSONAR
+public class CatalogSupplier implements CatalogMappingSupplier {
+
+    private static final String CUBE1 = "Cube1";
+    private static final String CUBE2 = "Cube2";
+    private static final String FACT = "Fact";
+
+    private static final String catalogBody = """
+            Cube with virtual cube with dimension references from other cubes and with calculatedMember
+            A virtual cube that combines measures and dimensions from multiple physical cubes into a unified analytical view.
+            Virtual cubes enable cross-cube analysis by creating a logical integration layer over existing physical cubes,
+            allowing users to analyze related metrics from different business processes in a single query
+            Catalog have two physical cubes Cube1 and Cube2 and virtual cube Cube1Cube2.
+            CalculatedMember from virtual cube uses MeasureCube1 from Cube1 and MeasureCube2 from Cube2 in formula.
+            """;
+
+    private static final String databaseSchemaBody = """
+            The cube defined in this example is based on a table that stores all the data.
+            The table is named `Fact` uses for Cube1 and contains two columns: `KEY` and `VALUE`.
+            The KEY column serves as a discriminator, while the `VALUE` column contains the measurements to be aggregated.
+            """;
+
+    private static final String queryBody = """
+            The bridge between the cube and the database is the query element. In this case, it is a TableQuery,
+            as it directly references the physical table `Fact`.
+            """;
+
+
+    private static final String measure1Body = """
+            Measure uses C1_Fact table VALUE column with sum aggregation in Cube1.
+    """;
+
+    private static final String measure2Body = """
+            Measure uses Fact table VALUE column with sum aggregation in Cube2.
+    """;
+
+    private static final String calculatedMemberBody = """
+            CalculatedMember uses MeasureCube1 from Cube1 and MeasureCube2 from Cube2 in formula.
+    """;
+
+    private static final String cube1Body = """
+            In this example uses cube with fact table Fact as data.
+            """;
+
+    private static final String cube2Body = """
+            In this example uses cube with fact table Fact as data.
+            """;
+
+    private static final String vCubeBody = """
+            Virtual cube uses mesures from Cube1 and Cube2. Virtual cube has references for them.
+            Also virtual cube has references to dimensions from Cube1 and Cube2
+            """;
+
+    private static final String schemaDocumentationTxt = """
+            A basic OLAP schema with virtual cube which have reference to Cube1, Cube2 with CalculatedMember
+            """;
+
+    @Override
+    public CatalogMapping get() {
+        DatabaseSchema databaseSchema = RolapMappingFactory.eINSTANCE.createDatabaseSchema();
+        databaseSchema.setId("databaseSchema");
+
+        Column keyColumn = RolapMappingFactory.eINSTANCE.createPhysicalColumn();
+        keyColumn.setName("KEY");
+        keyColumn.setId("_Fact_KEY");
+        keyColumn.setType(ColumnType.VARCHAR);
+
+        Column valueColumn = RolapMappingFactory.eINSTANCE.createPhysicalColumn();
+        valueColumn.setName("VALUE");
+        valueColumn.setId("_Fact_VALUE");
+        valueColumn.setType(ColumnType.INTEGER);
+
+        PhysicalTable table = RolapMappingFactory.eINSTANCE.createPhysicalTable();
+        table.setName(FACT);
+        table.setId("_Fact");
+        table.getColumns().addAll(List.of(keyColumn, valueColumn));
+        databaseSchema.getTables().add(table);
+
+        TableQuery query = RolapMappingFactory.eINSTANCE.createTableQuery();
+        query.setId("_FactQuery");
+        query.setTable(table);
+
+        SumMeasure measure1 = RolapMappingFactory.eINSTANCE.createSumMeasure();
+        measure1.setName("MeasureCube1");
+        measure1.setId("_MeasureCube1");
+        measure1.setColumn(valueColumn);
+
+        SumMeasure measure2 = RolapMappingFactory.eINSTANCE.createSumMeasure();
+        measure2.setName("MeasureCube2");
+        measure2.setId("_MeasureCube2");
+        measure2.setColumn(valueColumn);
+
+        MeasureGroup measureGroup1 = RolapMappingFactory.eINSTANCE.createMeasureGroup();
+        measureGroup1.getMeasures().add(measure1);
+
+        MeasureGroup measureGroup2 = RolapMappingFactory.eINSTANCE.createMeasureGroup();
+        measureGroup2.getMeasures().add(measure2);
+
+        Level level = RolapMappingFactory.eINSTANCE.createLevel();
+        level.setName("Level");
+        level.setId("_Level");
+        level.setColumn(keyColumn);
+
+        ExplicitHierarchy hierarchy = RolapMappingFactory.eINSTANCE.createExplicitHierarchy();
+        hierarchy.setHasAll(false);
+        hierarchy.setName("HierarchyWithoutHasAll");
+        hierarchy.setId("HierarchyWithoutHasAll");
+        hierarchy.setPrimaryKey(keyColumn);
+        hierarchy.setQuery(query);
+        hierarchy.getLevels().add(level);
+
+        StandardDimension dimension = RolapMappingFactory.eINSTANCE.createStandardDimension();
+        dimension.setName("Dimension1");
+        dimension.setId("_Dimension1");
+        dimension.getHierarchies().add(hierarchy);
+
+        DimensionConnector dimensionConnector1 = RolapMappingFactory.eINSTANCE.createDimensionConnector();
+        dimensionConnector1.setId("_dc_cube1Dimension1");
+        dimensionConnector1.setOverrideDimensionName("Cube1Dimension1");
+        dimensionConnector1.setForeignKey(keyColumn);
+        dimensionConnector1.setDimension(dimension);
+
+        DimensionConnector dimensionConnector2 = RolapMappingFactory.eINSTANCE.createDimensionConnector();
+        dimensionConnector2.setId("_dc_cube2Dimension1");
+        dimensionConnector2.setOverrideDimensionName("Cube2Dimension1");
+        dimensionConnector2.setForeignKey(keyColumn);
+        dimensionConnector2.setDimension(dimension);
+
+        PhysicalCube cube1 = RolapMappingFactory.eINSTANCE.createPhysicalCube();
+        cube1.setName(CUBE1);
+        cube1.setId("_Cube1");
+        cube1.setQuery(query);
+        cube1.getDimensionConnectors().add(dimensionConnector1);
+        cube1.getMeasureGroups().add(measureGroup1);
+        dimensionConnector1.setPhysicalCube(cube1);
+
+        PhysicalCube cube2 = RolapMappingFactory.eINSTANCE.createPhysicalCube();
+        cube2.setName(CUBE2);
+        cube2.setId("_Cube2");
+        cube2.setQuery(query);
+        cube2.getDimensionConnectors().add(dimensionConnector2);
+        cube2.getMeasureGroups().add(measureGroup2);
+        dimensionConnector2.setPhysicalCube(cube2);
+
+        CalculatedMember calculatedMember = RolapMappingFactory.eINSTANCE.createCalculatedMember();
+        calculatedMember.setName("Sum_Cub");
+        calculatedMember.setId("_cm_Sum_Cub");
+        calculatedMember.setHierarchy(hierarchy);
+        calculatedMember.setFormula("[Measures].[MeasureCube1] + [Measures].[MeasureCube2]");
+
+        VirtualCube vCube = RolapMappingFactory.eINSTANCE.createVirtualCube();
+        vCube.setName("Cube1Cube2");
+        vCube.setId("_Cube1Cube2");
+        vCube.getDimensionConnectors().addAll(List.of(dimensionConnector1, dimensionConnector2));
+        vCube.getReferencedMeasures().addAll(List.of(measure1, measure2));
+        vCube.getCalculatedMembers().add(calculatedMember);
+
+        Catalog catalog = RolapMappingFactory.eINSTANCE.createCatalog();
+        catalog.setName("Cube with virtual cube with calculatedMember");
+        catalog.setDescription("Schema with virtual cube with calculatedMember");
+        catalog.getCubes().addAll(List.of(cube1, cube2, vCube));
+        Documentation schemaDocumentation = RolapMappingFactory.eINSTANCE.createDocumentation();
+        schemaDocumentation.setValue(schemaDocumentationTxt);
+        catalog.getDocumentations().add(schemaDocumentation);
+        catalog.getDbschemas().add(databaseSchema);
+
+        document(catalog, "Schema with virtual cube with calculatedMember", catalogBody, 1, 0, 0, false, 0);
+        document(databaseSchema, "Database Schema", databaseSchemaBody, 1, 1, 0, true, 3);
+        document(query, "Query", queryBody, 1, 2, 0, true, 2);
+        document(measure1, "MeasureCube1", measure1Body, 1, 4, 0, true, 2);
+        document(measure2, "MeasureCube2", measure2Body, 1, 5, 0, true, 2);
+        document(calculatedMember, "Sum_Cub", calculatedMemberBody, 1, 5, 0, true, 2);
+        document(cube1, "Cube1", cube1Body, 1, 6, 0, true, 2);
+        document(cube1, "Cube1", cube2Body, 1, 7, 0, true, 2);
+        document(vCube, "VirtualCubeMeasureOnly", vCubeBody, 1, 7, 0, true, 2);
+
+        return catalog;
+    }
+
+}
