--- instance/emf/tutorial/virtualcubedimensions/src/main/java/org/eclipse/daanse/rolap/mapping/instance/emf/tutorial/virtualcubedimensions/CatalogSupplier.java
@@ -1,155 +1,0 @@
-/*
- * Copyright (c) 2025 Contributors to the Eclipse Foundation.
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *
- */
-package org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.virtualcubedimensions;
-
-import java.util.List;
-
-import org.eclipse.daanse.rolap.mapping.api.CatalogMappingSupplier;
-import org.eclipse.daanse.rolap.mapping.api.model.CatalogMapping;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Catalog;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Column;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.ColumnType;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.DatabaseSchema;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.DimensionConnector;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Documentation;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.ExplicitHierarchy;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Hierarchy;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.Level;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.MeasureGroup;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.PhysicalCube;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.PhysicalTable;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.RolapMappingFactory;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.StandardDimension;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.SumMeasure;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.TableQuery;
-import org.eclipse.daanse.rolap.mapping.emf.rolapmapping.VirtualCube;
-import org.osgi.service.component.annotations.Component;
-
-@Component(service = CatalogMappingSupplier.class)
-public class CatalogSupplier implements CatalogMappingSupplier {
-
-    private static final String CATALOG = "tutorial_20-01_VirtualCube_with_common_dimension";
-    private static final String CUBE1 = "Cube1";
-    private static final String CUBE2 = "Cube2";
-    private static final String FACT = "Fact";
-
-    private static final String schemaDocumentationTxt = """
-            A basic OLAP schema with virtual cube which have reference to Cube1, Cube2
-            """;
-
-    @Override
-    public CatalogMapping get() {
-        DatabaseSchema databaseSchema = RolapMappingFactory.eINSTANCE.createDatabaseSchema();
-        databaseSchema.setId("databaseSchema");
-
-        Column keyColumn = RolapMappingFactory.eINSTANCE.createPhysicalColumn();
-        keyColumn.setName("KEY");
-        keyColumn.setId("Fact_KEY");
-        keyColumn.setType(ColumnType.VARCHAR);
-
-        Column valueColumn = RolapMappingFactory.eINSTANCE.createPhysicalColumn();
-        valueColumn.setName("VALUE");
-        valueColumn.setId("Fact_VALUE");
-        valueColumn.setType(ColumnType.INTEGER);
-
-        PhysicalTable table = RolapMappingFactory.eINSTANCE.createPhysicalTable();
-        table.setName(FACT);
-        table.setId(FACT);
-        table.getColumns().addAll(List.of(keyColumn, valueColumn));
-        databaseSchema.getTables().add(table);
-
-        TableQuery query = RolapMappingFactory.eINSTANCE.createTableQuery();
-        query.setId("FactQuery");
-        query.setTable(table);
-
-        SumMeasure measure1 = RolapMappingFactory.eINSTANCE.createSumMeasure();
-        measure1.setName("MeasureCube1");
-        measure1.setId("MeasureCube1");
-        measure1.setColumn(valueColumn);
-
-        SumMeasure measure2 = RolapMappingFactory.eINSTANCE.createSumMeasure();
-        measure2.setName("MeasureCube2");
-        measure2.setId("MeasureCube2");
-        measure2.setColumn(valueColumn);
-
-        MeasureGroup measureGroup1 = RolapMappingFactory.eINSTANCE.createMeasureGroup();
-        measureGroup1.getMeasures().add(measure1);
-
-        MeasureGroup measureGroup2 = RolapMappingFactory.eINSTANCE.createMeasureGroup();
-        measureGroup2.getMeasures().add(measure2);
-
-        Level level = RolapMappingFactory.eINSTANCE.createLevel();
-        level.setName("Level2");
-        level.setId("Level2");
-        level.setColumn(keyColumn);
-
-        ExplicitHierarchy hierarchy = RolapMappingFactory.eINSTANCE.createExplicitHierarchy();
-        hierarchy.setHasAll(false);
-        hierarchy.setName("HierarchyWithoutHasAll");
-        hierarchy.setId("HierarchyWithoutHasAll");
-        hierarchy.setPrimaryKey(keyColumn);
-        hierarchy.setQuery(query);
-        hierarchy.getLevels().add(level);
-
-        StandardDimension dimension = RolapMappingFactory.eINSTANCE.createStandardDimension();
-        dimension.setName("Dimension1");
-        dimension.setId("Dimension1");
-        dimension.getHierarchies().add(hierarchy);
-
-        DimensionConnector dimensionConnector1 = RolapMappingFactory.eINSTANCE.createDimensionConnector();
-        dimensionConnector1.setId("_dc_cube1Dimension1");
-        dimensionConnector1.setOverrideDimensionName("Cube1Dimension1");
-        dimensionConnector1.setForeignKey(keyColumn);
-        dimensionConnector1.setDimension(dimension);
-
-        DimensionConnector dimensionConnector2 = RolapMappingFactory.eINSTANCE.createDimensionConnector();
-        dimensionConnector2.setId("_dc_cube2Dimension1");
-        dimensionConnector2.setOverrideDimensionName("Cube2Dimension1");
-        dimensionConnector2.setForeignKey(keyColumn);
-        dimensionConnector2.setDimension(dimension);
-
-        PhysicalCube cube1 = RolapMappingFactory.eINSTANCE.createPhysicalCube();
-        cube1.setName(CUBE1);
-        cube1.setId(CUBE1);
-        cube1.setQuery(query);
-        cube1.getDimensionConnectors().add(dimensionConnector1);
-        cube1.getMeasureGroups().add(measureGroup1);
-        dimensionConnector1.setPhysicalCube(cube1);
-
-        PhysicalCube cube2 = RolapMappingFactory.eINSTANCE.createPhysicalCube();
-        cube2.setName(CUBE2);
-        cube2.setId(CUBE2);
-        cube2.setQuery(query);
-        cube2.getDimensionConnectors().add(dimensionConnector2);
-        cube2.getMeasureGroups().add(measureGroup2);
-        dimensionConnector2.setPhysicalCube(cube2);
-
-        VirtualCube vCube = RolapMappingFactory.eINSTANCE.createVirtualCube();
-        vCube.setName("Cube1Cube2");
-        vCube.setId("Cube1Cube2");
-        vCube.getDimensionConnectors().addAll(List.of(dimensionConnector1, dimensionConnector2));
-        vCube.getReferencedMeasures().addAll(List.of(measure1, measure2));
-
-        Catalog catalog = RolapMappingFactory.eINSTANCE.createCatalog();
-        catalog.setName("Cube_with_virtual_cube");
-        catalog.setDescription("Schema with virtual cube");
-        catalog.getCubes().addAll(List.of(cube1, cube2, vCube));
-        Documentation schemaDocumentation = RolapMappingFactory.eINSTANCE.createDocumentation();
-        schemaDocumentation.setValue(schemaDocumentationTxt);
-        catalog.getDocumentations().add(schemaDocumentation);
-        catalog.getDbschemas().add(databaseSchema);
-
-        return catalog;
-    }
-
-}
