--- instance/emf/tutorial/virtualcubemeasureonly/src/main/resources/data/C2_Fact.csv
@@ -1,3 +1,0 @@
-"KEY","VALUE"
-VARCHAR,INTEGER
-C,82
