--- instance/emf/tutorial/virtualcubevisiblereferencecubes/src/main/resources/data/Fact.csv
@@ -1,4 +1,0 @@
-"KEY","VALUE"
-VARCHAR,INTEGER
-A,42
-B,21
